Added recovery mode for shield-agent for emergency restores
To aid in restoring things like BOSH + SHIELD databases in an
emergency situation, an emergency-recovery script was added, which
uses properties in `shield.agent.recovery.*` and the `store_key` of
the archive to restore, provided by the user.

This is intended as a manual emergency restore. You deploy BOSH,
or SHIELD, and ssh into the node, then run this script manually,
providing the archive's `store_key` as the only argument.

The script will then run `shield-pipe`, provided things have been
configured properly, and it will restore your data from backup,
without the need for a working SHIELD installation that knows about
all of your archives.

If it wasn't clear already, this script is intended to be used
IN EMERGENCY SITUATIONS ONLY, to RESTORE data.