Ensure that the postgres plugin finds the bin_dir by default
The shield postgres plugin expects the bin_dir to be located at
/var/vcap/packages/postgres/bin. The shield-boshrelease agent-pgtools
job will put the bin_dir to /var/vcap/packages/postgres9.4/bin.

The change in the init script will check if the default bin_dir
exists and if not, it will create a symlink to the one of the
shield-boshrelease agent-pgtools.