Handle Credentials better
If you don't have credhub, and just want to demo / eval, you can deploy
with a `-l manifest/creds.yml`, at least for the next 100 years or so.

If you forget to generate the creds, or in any other way equip BOSH to
generate the creds for you, the template will error out, instead of
attempting to deploy with '((placeholder))' credentials.