Use Full Credentials in errand import
This commit introduces the ability to (insecurely) use your full SHIELD
credentials in the errand import, instead of a risk-mitigated auth
token.  If you set your username and password, the import errand will
use them to log into SHIELD via the CLI, generate an auth token, use it
for the import, and then revoke it when it is finished.

This is insecure, since you are giving out credentials that may be
difficult to change, vs. an auth token which can be revoked.

Fixes #114 and closes #118